add track clear cart and id abandoned cart fix

+++ Teavaro Ecomm Demo App/Teavaro Ecomm Demo App/Data/CoreData/AbandonedCarts+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension AbandonedCarts {
         return NSFetchRequest<AbandonedCarts>(entityName: "AbandonedCarts")
     }
 
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int
     @NSManaged public var items: NSSet?
 
 }
